Changing ports, adding active_players
To run API as back end to other flask apps there needs to not be port collision.  Changed flask to running on port 8888 and the nginx proxy on port 8080.

API now accessed via http://localhost:8080/

Other apps can be run at the same time using http://localhost/

Need way to quickly grab players that are active so that bit was added.